fix(html-slides): normalize runtime script sources
