+++ src/lib/htmlSlides/nativeHtmlSlides.ts
@@ -1,5 +1,6 @@
 import type { Slide, SlideImageLicenseStrictness, SlideImageRole, SlideImageSourcePreference } from '../markdownToSlides';
 import type { ResolvedSlideAsset, SlideAssetRecord, SlideImageIntent } from '../../services/slideAssets';
+import { hasDangerousUrlScheme, normalizeRuntimeScriptSrc } from './htmlUrlSafety';
 
 export interface HtmlNativeAssetIntentInput {
   id?: unknown;
@@ -90,56 +91,8 @@ function scriptSrcs(html: string): string[] {
     .filter((src): src is string => Boolean(src));
 }
 
-const HTML_ENTITY_MAP: Record<string, string> = {
-  amp: '&',
-  apos: "'",
-  colon: ':',
-  gt: '>',
-  lt: '<',
-  NewLine: '\n',
-  quot: '"',
-  Tab: '\t',
-};
-
-function decodeHtmlAttributeValue(value: string): string {
-  const codePointToString = (codePoint: number, fallback: string) =>
-    Number.isInteger(codePoint) && codePoint >= 0 && codePoint <= 0x10ffff
-      ? String.fromCodePoint(codePoint)
-      : fallback;
-  return value.replace(/&(#x[0-9a-f]+;?|#\d+;?|[a-z][a-z0-9]+;?)/gi, (entity, body: string) => {
-    const clean = body.endsWith(';') ? body.slice(0, -1) : body;
-    if (clean.toLowerCase().startsWith('#x')) {
-      const codePoint = Number.parseInt(clean.slice(2), 16);
-      return Number.isFinite(codePoint) ? codePointToString(codePoint, entity) : entity;
-    }
-    if (clean.startsWith('#')) {
-      const codePoint = Number.parseInt(clean.slice(1), 10);
-      return Number.isFinite(codePoint) ? codePointToString(codePoint, entity) : entity;
-    }
-    return HTML_ENTITY_MAP[clean] ?? entity;
-  });
-}
-
-function unquoteAttributeValue(raw: string): string {
-  const trimmed = raw.trim();
-  const quote = trimmed[0];
-  if ((quote === '"' || quote === "'") && trimmed.endsWith(quote)) return trimmed.slice(1, -1);
-  return trimmed;
-}
-
-function normalizedUrlForSchemeCheck(rawValue: string): string {
-  return decodeHtmlAttributeValue(unquoteAttributeValue(rawValue))
-    .replace(/[\u0000-\u001f\u007f\s]+/g, '')
-    .toLowerCase();
-}
-
-function hasDangerousUrlScheme(rawValue: string): boolean {
-  const normalized = normalizedUrlForSchemeCheck(rawValue);
-  return normalized.startsWith('javascript:') || normalized.startsWith('vbscript:');
-}
-
 function isAllowedTemplateRuntimeScript(src: string): boolean {
-  const clean = decodeHtmlAttributeValue(src).split(/[?#]/, 1)[0]?.trim().replace(/\\/g, '/') ?? '';
+  const clean = normalizeRuntimeScriptSrc(src);
   if (!clean) return false;
   if (/^https:\/\/cdn\.jsdelivr\.net\/npm\/chart\.js(?:@[\w.-]+)?(?:\/[^?#]*)?$/i.test(clean)) return true;
   if (/^(?:[a-z][a-z0-9+.-]*:|\/\/|\/)/i.test(clean)) return false;
